move from_csv specific sources tests into csv test module
